+++ CHANGELOG.md
@@ -1,3 +1,20 @@
--   **1.0.0**
+## v1.0.0
 
-    -   Mod Release
+  - Mod Release
+
+## v1.0.1
+
+  - Fix missing Lethalthings
+
+## v2.0.0
+
+  - Nothing
+
+## v2.0.1
+
+  - Separate witch sound for discovering deadbody
+
+## v2.0.2
+
+  - Allow config
+  - Subaru's phone ring for icecream truck delivery sound